Add player orientation
This is mostly experimental and is a little messy. Concerns are mixed
up, some code is redundant. I decided to let things like that because I
don't know how to refactor it cleanly enough so that we don't have to do it again
later, I am hoping that by adding things organically we start seeing the
redundant patterns and that it will help us deciding on the best
organization.